Split out disassemble to remove numeric from instruction includes.
